bitguard-src/requirements.txt
fastapi
uvicorn[standard]
cryptography
argon2-cffi
sqlalchemy
psycopg2-binary
python-dotenv
